+++ src/cjs/utils/detectDarkMode.d.ts
@@ -1,0 +1,10 @@
+/**
+ * @author  Ikaros Kappler
+ * @date    2021-11-07
+ * @modified 2023-09-25 Ported to Typescript and added system darkmode detection.
+ * @version 1.1.0
+ *
+ * @param {Record<string,string>} GUP
+ * @returns {boolean}
+ */
+export declare const detectDarkMode: (GUP?: Record<string, string>) => any;