+++ src/ts/Polygon.ts
@@ -24,7 +24,9 @@
  * @modified 2022-02-02 Added the `destroy` method.
  * @modified 2022-02-02 Cleared the `Polygon.toSVGString` function (deprecated). Use `drawutilssvg` instead.
  * @modified 2022-03-08 Added the `Polygon.clone()` function.
- * @version 1.10.0
+ * @modified 2023-09-25 Added the `Polygon.getInterpolationPolygon(number)` function.
+ * @modified 2023-09-25 Added the `Polygon.lineIntersections(Line)` function.
+ * @version 1.11.0
  *
  * @file Polygon
  * @public
@@ -32,7 +34,9 @@
 
 import { BezierPath } from "./BezierPath";
 import { Bounds } from "./Bounds";
+import { Line } from "./Line";
 import { UIDGenerator } from "./UIDGenerator";
+import { VertTuple } from "./VertTuple";
 import { Vertex } from "./Vertex";
 import { XYCoords, SVGSerializable, UID } from "./interfaces";
 
@@ -274,6 +278,75 @@ export class Polygon implements SVGSerializable {
       this.vertices[i].rotate(angle, center);
     }
     return this;
+  }
+
+  /**
+   * Get all line intersections with this polygon.
+   *
+   * See demo `47-closest-vector-projection-on-polygon` for how it works.
+   *
+   * @param {VertTuple} line - The line to find intersections with (no bounds, just an infinite line).
+   * @returns {Array<Vertex>} - An array of all intersections within the polygon bounds.
+   */
+  lineIntersections(line: VertTuple<any>): Array<Vertex> {
+    // Find the intersections of all lines inside the edge bounds
+    const intersectionPoints: Array<Vertex> = [];
+    for (var i = 0; i < this.vertices.length; i++) {
+      const polyLine = new Line(this.vertices[i], this.vertices[(i + 1) % this.vertices.length]);
+      const intersection = polyLine.intersection(line);
+      // true => only inside bounds
+      // ignore last edge if open
+      if ((!this.isOpen || i + 1 !== this.vertices.length) && intersection !== null && polyLine.hasPoint(intersection, true)) {
+        intersectionPoints.push(intersection);
+      }
+    }
+    return intersectionPoints;
+  }
+
+  closestLinetIntersection(line: VertTuple<any>): Vertex | null {
+    const allIntersections = this.lineIntersections(line);
+    if (allIntersections.length <= 0) {
+      // Empty polygon -> no intersections
+      return null;
+    }
+    // Find the closest intersection
+    let closestIntersection = new Vertex(Number.MAX_VALUE, Number.MAX_VALUE);
+    let curDist = Number.MAX_VALUE;
+    for (var i in allIntersections) {
+      const curVert = allIntersections[i];
+      const dist = curVert.distance(line.a);
+      if (dist < curDist && line.hasPoint(curVert)) {
+        curDist = dist;
+        closestIntersection = curVert;
+      }
+    }
+    return closestIntersection;
+  }
+
+  /**
+   * Construct a new polygon from this polygon with more vertices on each edge. The
+   * interpolation count determines the number of additional vertices on each edge.
+   * An interpolation count of `0` will return a polygon that equals the source
+   * polygon.
+   *
+   * @param {number} interpolationCount
+   * @returns {Polygon} A polygon with `interpolationCount` more vertices (as as factor).
+   */
+  getInterpolationPolygon(interpolationCount: number): Polygon {
+    const verts: Array<Vertex> = [];
+    for (var i = 0; i < this.vertices.length; i++) {
+      const curVert = this.vertices[i];
+      const nextVert = this.vertices[(i + 1) % this.vertices.length];
+      verts.push(curVert.clone());
+      // Add interpolation points
+      if (!this.isOpen || i + 1 !== this.vertices.length) {
+        const lerpAmount = 1.0 / (interpolationCount + 1);
+        for (var j = 1; j <= interpolationCount; j++) {
+          verts.push(curVert.clone().lerp(nextVert, lerpAmount * j));
+        }
+      }
+    }
+    return new Polygon(verts, this.isOpen);
   }
 
   /**