+++ changelog.md
@@ -1,5 +1,13 @@
 # Changelog
 
+- 2023-09-25
+
+  - Added the `Polygon.getInterpolationPolygon(number)` function.
+  - Added the `detectDarkMode` helper function.
+  - Adding the demo `47-closest-vector-projection-on-polygon`.
+  - Changed param type of `Line.intersection()` from Line to VertTuple.
+  - Added `Polyon.lineIntersections(VertTuple)` function.
+
 - 2023-03-06
   - Finalizing the audio demo (demo-46).
 - 2023-02-10